Env.Internal.Help

Documentation

helpInfo :: Info e -> Parser e b -> [(String, e)] -> String

helpDoc :: Parser e a -> String

A pretty-printed list of recognized environment variables suitable for usage messages

data Info e

Parser's metadata

type ErrorHandler e = String -> e -> Maybe String

Given a variable name and an error value, try to produce a useful error message

defaultInfo :: Info Error

defaultErrorHandler :: (AsUnset e, AsEmpty e, AsUnread e) => ErrorHandler e

The default error handler

header :: String -> Info e -> Info e

Set the help text header (it usually includes the application's name and version)

desc :: String -> Info e -> Info e

Set the short description

footer :: String -> Info e -> Info e

Set the help text footer (it usually includes examples)

handleError :: ErrorHandler e -> Info x -> Info e

An error handler